Ensure ENV variables can support booleans
Contradictory to what a shell ENV variable type

This is to allign us more with what we set in the `values.yml` file.
We know that helm supports `boolean` types in the `values.yml`, so
let´s allow the user to be able to set this values, throught the
ENV vars cmd.